Update: 增强持久层，SqliteDataSource不使用C3P0连接池。 Fixed: 修复DataSourceUtil中资源释放机制缺失。

+++ persistence/src/main/java/psn/ifplusor/persistence/entity/EntityUtil.java
@@ -292,7 +292,7 @@ public class EntityUtil {
 				} else if (type == Long.class || type == long.class) {
 					property.setter.invoke(obj, rs.getLong(column));
 				} else if (type == Float.class || type == float.class) {
-					property.setter.invoke(obj, rs.getShort(column));
+					property.setter.invoke(obj, rs.getFloat(column));
 				} else if (type == Double.class || type == double.class) {
 					property.setter.invoke(obj, rs.getDouble(column));
 				} else if (type == byte[].class) {
